refactor: deleted unnecessary components, some explaining comments added
